+++ src/components/film-extra.js
@@ -1,0 +1,9 @@
+const createFilmExtraTemplate = () => {
+  return (
+    `<section class="films-list--extra">
+      <h2 class="films-list__title">Top rated</h2>
+    </section>`
+  );
+};
+
+export {createFilmExtraTemplate};
