+++ src/components/movie-container.js
@@ -1,0 +1,7 @@
+const createMovieContainerTemplate = () => {
+  return (
+    `<div class="films-list__container"></div>`
+  );
+};
+
+export {createMovieContainerTemplate};
