+++ src/components/movie-list.js
@@ -1,0 +1,9 @@
+const createMovieListTemplate = () => {
+  return (
+    `<section class="films-list">
+      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+    </section>`
+  );
+};
+
+export {createMovieListTemplate};
