+++ src/components/movie-main.js
@@ -1,0 +1,7 @@
+const createMovieMainTemplate = () => {
+  return (
+    `<section class="films"></section>`
+  );
+};
+
+export {createMovieMainTemplate};
