+++ src/const.js
@@ -1,0 +1,4 @@
+const MOVIE_COUNT = 5;
+const EXTRA_MOVIE_COUNT = 2;
+
+export {MOVIE_COUNT, EXTRA_MOVIE_COUNT};
